Simplify Quarry creation
