+++ programs/quarry-mine/src/instructions/mod.rs
@@ -1,7 +1,11 @@
 pub mod claim_rewards;
 pub mod claim_rewards_v2;
+pub mod create_quarry;
+pub mod create_quarry_v2;
 pub mod rescue_tokens;
 
 pub use claim_rewards::*;
 pub use claim_rewards_v2::*;
+pub use create_quarry::*;
+pub use create_quarry_v2::*;
 pub use rescue_tokens::*;
